# apps/backend/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "User": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "reset_password_token": {
            "type": "string"
          },
          "email_verification_timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "blacklisted_tokens": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": ["id", "email"],
        "type": "object",
        "additionalProperties": false
      },
      "ValidateErrorJSON": {
        "properties": {
          "message": {
            "type": "string",
            "enum": ["Validation failed"],
            "nullable": false
          },
          "details": {
            "properties": {},
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": ["message", "details"],
        "type": "object",
        "additionalProperties": false
      },
      "UserCreationBody": {
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"],
        "type": "object",
        "additionalProperties": false
      },
      "Todo": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "todo": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "todo"],
        "type": "object",
        "additionalProperties": false
      },
      "TodoRequestBody": {
        "properties": {
          "todo": {
            "type": "string"
          }
        },
        "required": ["todo"],
        "type": "object",
        "additionalProperties": false
      },
      "UserLoginBody": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"],
        "type": "object",
        "additionalProperties": false
      },
      "FilteredUserInterface": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": ["id", "name", "email"],
        "type": "object",
        "additionalProperties": false
      },
      "ConfirmResetPasswordBody": {
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirm_password": {
            "type": "string"
          }
        },
        "required": ["token", "password", "confirm_password"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "OAuth2PasswordSecurity": {
        "type": "oauth2",
        "description": "OAuth2 Password Grant",
        "flows": {
          "password": {
            "tokenUrl": "http://localhost:3000/api/v1/auth/login",
            "scopes": {}
          }
        }
      },
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token only"
      }
    }
  },
  "info": {
    "title": "@myorganizer/source",
    "version": "0.0.0",
    "license": {
      "name": "MIT"
    },
    "contact": {}
  },
  "paths": {
    "/user": {
      "get": {
        "operationId": "GetAllUsers",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["Users Management"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": []
      },
      "post": {
        "operationId": "CreateUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {}
                  ]
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateErrorJSON"
                }
              }
            }
          }
        },
        "tags": ["Users Management"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreationBody"
              }
            }
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "operationId": "GetUserById",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    }
                  ],
                  "nullable": true
                }
              }
            }
          }
        },
        "tags": ["Users Management"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/todo": {
      "get": {
        "operationId": "GetAllTodos",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["Todo Management"],
        "security": [],
        "parameters": []
      },
      "post": {
        "operationId": "CreateTodo",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          }
        },
        "tags": ["Todo Management"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoRequestBody"
              }
            }
          }
        }
      }
    },
    "/todo/{todoId}": {
      "delete": {
        "operationId": "DeleteTodo",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "tags": ["Todo Management"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "todoId",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status_code": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["message", "status_code"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Authentication"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginBody"
              }
            }
          }
        }
      }
    },
    "/auth/logout/{userId}": {
      "post": {
        "operationId": "Logout",
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["message", "status"],
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Failed to logout"
          }
        },
        "tags": ["Authentication"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "RefreshToken",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["user", "message", "status"],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateErrorJSON"
                }
              }
            }
          }
        },
        "tags": ["Authentication"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                },
                "required": ["refresh_token"],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "CreateUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {}
                  ]
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateErrorJSON"
                }
              }
            }
          }
        },
        "tags": ["Authentication"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreationBody"
              }
            }
          }
        }
      }
    },
    "/auth/verify/email": {
      "patch": {
        "operationId": "VerifyEmail",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredUserInterface"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateErrorJSON"
                }
              }
            }
          }
        },
        "tags": ["Authentication"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "token": {
                    "type": "string"
                  }
                },
                "required": ["token"],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/auth/verify/resend/{userId}": {
      "post": {
        "operationId": "ResendVerificationEmail",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "tags": ["Authentication"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/password/reset": {
      "post": {
        "operationId": "ResetPassword",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["message", "status"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Authentication"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": ["email"],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/auth/password/reset/confirm": {
      "patch": {
        "operationId": "ConfirmResetPassword",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["message", "status"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Authentication"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmResetPasswordBody"
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1/",
      "description": "Local server"
    }
  ]
}